+++ src/client/common/terminal/shellDetectors/settingsShellDetector.ts
@@ -1,0 +1,62 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+
+import { inject, injectable } from 'inversify';
+import { Terminal } from 'vscode';
+import { IWorkspaceService } from '../../application/types';
+import { traceVerbose } from '../../logger';
+import { IPlatformService } from '../../platform/types';
+import { OSType } from '../../utils/platform';
+import { ShellIdentificationTelemetry, TerminalShellType } from '../types';
+import { BaseShellDetector } from './baseShellDetector';
+
+/**
+ * Identifies the shell based on the user settings.
+ *
+ * @export
+ * @class SettingsShellDetector
+ * @extends {BaseShellDetector}
+ */
+@injectable()
+export class SettingsShellDetector extends BaseShellDetector {
+    constructor(@inject(IWorkspaceService) private readonly workspace: IWorkspaceService,
+        @inject(IPlatformService) private readonly platform: IPlatformService) {
+        super(1);
+    }
+    public getTerminalShellPath(): string | undefined {
+        const shellConfig = this.workspace.getConfiguration('terminal.integrated.shell');
+        let osSection = '';
+        switch (this.platform.osType) {
+            case OSType.Windows: {
+                osSection = 'windows';
+                break;
+            }
+            case OSType.OSX: {
+                osSection = 'osx';
+                break;
+            }
+            case OSType.Linux: {
+                osSection = 'linux';
+                break;
+            }
+            default: {
+                return '';
+            }
+        }
+        return shellConfig.get<string>(osSection)!;
+    }
+    public identify(telemetryProperties: ShellIdentificationTelemetry, _terminal?: Terminal): TerminalShellType | undefined {
+        const shellPath = this.getTerminalShellPath();
+        telemetryProperties.hasCustomShell = !!shellPath;
+        const shell = shellPath ? this.identifyShellFromShellPath(shellPath) : TerminalShellType.other;
+
+        if (shell !== TerminalShellType.other) {
+            telemetryProperties.shellIdentificationSource = 'environment';
+        }
+        telemetryProperties.shellIdentificationSource = 'settings';
+        traceVerbose(`Shell path from user settings '${shellPath}'`);
+        return shell;
+    }
+}
