+++ src/test/common/serviceRegistry.unit.test.ts
@@ -43,7 +43,9 @@ import { PipEnvActivationCommandProvider } from '../../client/common/terminal/en
 import { PyEnvActivationCommandProvider } from '../../client/common/terminal/environmentActivationProviders/pyenvActivationProvider';
 import { TerminalServiceFactory } from '../../client/common/terminal/factory';
 import { TerminalHelper } from '../../client/common/terminal/helper';
-import { SettingsShellDetector, TerminalNameShellDetector, UserEnvironmentShellDetector } from '../../client/common/terminal/shellDetectors';
+import { SettingsShellDetector } from '../../client/common/terminal/shellDetectors/settingsShellDetector';
+import { TerminalNameShellDetector } from '../../client/common/terminal/shellDetectors/terminalNameShellDetector';
+import { UserEnvironmentShellDetector } from '../../client/common/terminal/shellDetectors/userEnvironmentShellDetector';
 import { IShellDetector, ITerminalActivationCommandProvider, ITerminalActivationHandler, ITerminalActivator, ITerminalHelper, ITerminalServiceFactory, TerminalActivationProviders } from '../../client/common/terminal/types';
 import { IAsyncDisposableRegistry, IBrowserService, IConfigurationService, ICryptoUtils, ICurrentProcess, IEditorUtils, IExperimentsManager, IExtensions, IFeatureDeprecationManager, IHttpClient, IInstaller, ILogger, IPathUtils, IPersistentStateFactory, IRandom } from '../../client/common/types';
 import { IMultiStepInputFactory, MultiStepInputFactory } from '../../client/common/utils/multiStepInput';
