+++ src/test/common/terminals/helper.unit.test.ts
@@ -24,12 +24,12 @@ import {
 } from '../../../client/common/terminal/environmentActivationProviders/pyenvActivationProvider';
 import { TerminalHelper } from '../../../client/common/terminal/helper';
 import { ShellDetector } from '../../../client/common/terminal/shellDetector';
-import { TerminalNameShellDetector } from '../../../client/common/terminal/shellDetectors';
+import { TerminalNameShellDetector } from '../../../client/common/terminal/shellDetectors/terminalNameShellDetector';
 import {
+    IShellDetector,
     ITerminalActivationCommandProvider,
     TerminalShellType
 } from '../../../client/common/terminal/types';
-import { IShellDetector } from '../../../client/common/terminal/types';
 import { IConfigurationService } from '../../../client/common/types';
 import { getNamesAndValues } from '../../../client/common/utils/enum';
 import { Architecture, OSType } from '../../../client/common/utils/platform';
